requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
greenlet==3.0.3
humanize==4.10.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
MouseInfo==0.1.3
phonenumbers==8.13.37
pillow==10.3.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.9.0
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
pywhatkit==5.4
pywin32==306
qrcode==7.4.2
razorpay==1.4.2
redis==5.0.7
reportlab==4.2.2
requests==2.32.3
setuptools==70.3.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wikipedia==1.4.0
